requirements_20250813.txt
biopython==1.85
cutadapt==5.0
dnaio==1.2.3
edlib==1.3.9.post1
isal==1.7.2
llvmlite==0.44.0
numba==0.61.2
numpy==2.2.6
parasail==1.3.4
pysam==0.23.3
tqdm==4.67.1
xopen==2.0.2
zlib-ng==0.5.1
